requirements.txt
fastapi==0.111.0
tensorflow==2.13.0
numpy==1.24.3 
opencv-python==4.10.0.84
pillow==9.4.0
